% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.lib.R
\name{AnalyseStableModels_LOO}
\alias{AnalyseStableModels_LOO}
\title{analyse stability of models from digest}
\usage{
AnalyseStableModels_LOO(X, y, clf, tmp, loo)
}
\arguments{
\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}

\item{tmp:}{the digested result object from digest}
}
\value{
a list of each sparsity the frequency of each feature of empirical best model in k-folds cross validation
}
\description{
This function analyses prevalence of features of best model of different sparsity in crossval (here still k-folds)
}
