% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{analyzePopulationFeatures}
\alias{analyzePopulationFeatures}
\title{Prints as text the detail on a given experiment along with summarized results (if computed)}
\usage{
analyzePopulationFeatures(
  pop,
  X,
  y,
  res_clf,
  makeplot = TRUE,
  name = "",
  ord.feat = "importance",
  make.network = TRUE,
  network.layout = "circular",
  network.alpha = 1e-04,
  verbose = TRUE,
  pdf.dims = c(width = 25, height = 20),
  filter.perc = 0.05,
  k_penalty = 0.75/100,
  k_max = 0
)
}
\arguments{
\item{pop:}{a population of models}

\item{X:}{the X dataset where to compute the abundance and prevalence}

\item{y:}{the target class}

\item{res_clf:}{the results of the classifier as well as the config object}

\item{makeplot:}{make a pdf file with the resulting plots (default:TRUE)}

\item{name:}{the suffix of the pdf file (default:"")}

\item{ord.feat:}{which ordering approch to use for the features (default:importance) in the models, anything 
else will compute automatic hierarchical ordering based on the manhattan distance}

\item{make.network:}{build a network and print it out in the pdf}

\item{network.layout:}{the network layout by default is circular (layout_in_circle) and will be a weighted Fruchterman-Reingold otherwise}

\item{network.alpha:}{threshold of significance for the network (default:1e-4)}

\item{verbose:}{print out informaiton}

\item{pdf.dims:}{dimensions of the pdf object (default: c(w = 25, h = 20))}

\item{filter.perc:}{filter by prevalence percentage in the population between 0 and 1 (default:0.05)}

\item{k_penalty:}{the sparsity penalty needed to select the best models of the population (default:0.75/100).}

\item{k_max:}{select the best population below a given threshold. If (default:0) no selection is performed.}
}
\value{
plots if makeplot is FALSE
}
\description{
This function takes a population of models and makes three plots, feature prevalence in population, 
feature abundance by class and feature prevalence by class
}
