% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.lib.R
\name{bestModelStability}
\alias{bestModelStability}
\title{analyse stability of models from digest}
\usage{
bestModelStability(X, y, clf, digested.result, method = "exact")
}
\arguments{
\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}

\item{digested.result:}{the digest result from digest}

\item{method:}{wether to compute the stability of the best compared to the best in the folds (exact), or the top best (fuzzy)}
}
\value{
an object with first a list of feature presence tables for each k_sparsity and a list of feature presence frequency
}
\description{
This function analyses prevalence of features of best model of different sparsity in crossval (here still k-folds)
}
