% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cir_test}
\alias{cir_test}
\title{Cirhosis stage 2 (frequencies)}
\description{
This dataset consists of frequency abundance files as downloaded from http://waldronlab.io/curatedMetagenomicData/
This is a list containing two elements: (i) the X data matrix with 1045 species and 56 observations and (ii) patient class = -1 (n=25) and healthy controls (n=31)
}
\author{
Qin, Nan, Fengling Yang, Ang Li, Edi Prifti, Yanfei Chen, Li Shao, Jing Guo, et al “Alterations of the human gut microbiome in liver cirrhosis.” Nature 513, no. 7516 (July 23, 2014): 59–64.
}
\keyword{cirrhosis,}
\keyword{liver}
\keyword{microbiome,}
\keyword{species}
