% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cir_train}
\alias{cir_train}
\title{Cirhosis stage 1 (frequencies)}
\description{
This dataset consists of frequency abundance files as downloaded from http://waldronlab.io/curatedMetagenomicData/. 
This is a list containing two elements: (i) the X data matrix with 1045 species and 181 observations and (ii) patient class = -1 (n=98) and healthy controls (n=83)
}
\author{
Qin, Nan, Fengling Yang, Ang Li, Edi Prifti, Yanfei Chen, Li Shao, Jing Guo, et al “Alterations of the human gut microbiome in liver cirrhosis.” Nature 513, no. 7516 (July 23, 2014): 59–64 _.
}
\keyword{cirrhosis,}
\keyword{liver}
\keyword{microbiome,}
\keyword{species}
