% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{cleanPopulation}
\alias{cleanPopulation}
\title{cleanPopulation}
\usage{
cleanPopulation(pop, clf)
}
\arguments{
\item{pop:}{is population (a list) of predomics objects}

\item{clf:}{the classifier object}
}
\value{
a population of predomics objects
}
\description{
Looks for invalid predomics objects in a population and removes them.
}
