% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{computeCardEnrichment}
\alias{computeCardEnrichment}
\title{computeCardEnrichment}
\usage{
computeCardEnrichment(v.card.mat, y)
}
\arguments{
\item{v.card.mat:}{a dataframe with the cardinality of each feature (columns) and each group in the y vector (rows)}

\item{y:}{the vector containing the class specification for each sample}
}
\value{
a data.frame with the statistics computed
}
\description{
Computes statistic for enrichment of the cardinality of a score for a two class vector
}
