% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{computeCoeffSVMLin}
\alias{computeCoeffSVMLin}
\title{Compute other prediction scores such as precision, recall and f-score}
\usage{
computeCoeffSVMLin(X, y, clf = NULL, mod = NULL)
}
\arguments{
\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{mod:}{a predomics object to be updated}

\item{clf:}{an object containing the different parameters of the classifier}
}
\value{
a model whose evaluation parameters are updated or a list containing coefficients and intercept if mod is not set.
}
\description{
This function computes prediction scores based on the confusion matrix such as accuracy, precision, recall and f-score
}
