% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{computeConfusionMatrix}
\alias{computeConfusionMatrix}
\title{Evaluates the confusion Matrix of the predicted class and the class to predict}
\usage{
computeConfusionMatrix(mod, X, y, clf)
}
\arguments{
\item{mod:}{a model object to be evaluated}

\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}
}
\value{
a confusion matrix
}
\description{
This function evaluates the accuracy of a model
}
