% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeImportantFeaturesFBM.R
\name{computeEffectSizes}
\alias{computeEffectSizes}
\title{Compute effect sizes for features in binary classification/regression tasks}
\usage{
computeEffectSizes(X, y, mode)
}
\arguments{
\item{X}{The X matrix (rows=features; columns=samples)}

\item{y}{The y vector of sample class (-1,1 in binary classification; 
continuous variable in regression)}

\item{mode}{classification or regression}
}
\value{
data frame of features, effect sizes (cliff's delta for binary 
classification; spearman rho for regression), and pvalues (wicoxon rank-sum 
test for binary classification task; spearman correlation for regression)
}
\description{
In binary classification tasks, compute the cliff's delta effect 
sizes btw groups (1 vs. -1) + pvalues from wilcoxon rank-sum tests; in 
regression tasks, compute spearman correlations (rho + pvalue) vs. 
continuous y variable
}
