% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{computeFeatureMetrics}
\alias{computeFeatureMetrics}
\title{Computes different metrics for a given distributions}
\usage{
computeFeatureMetrics(data)
}
\arguments{
\item{data:}{a data frame containing the data to be treated.}
}
\value{
a data frame containing different metrics: variance_to_mean, signal_to_noise, variation_coefficient, efficiency and quartile_dispertion
}
\description{
This function computes to compute a certain number of metrics on a dataset for each variable 
(rows, such as prevalence, quartile distribution, etc.)
}
