% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{computeIntercept}
\alias{computeIntercept}
\title{Computes the best intercept for the model while minimizing error}
\usage{
computeIntercept(score, y, verbose = FALSE, sign = "auto", plot = FALSE)
}
\arguments{
\item{score:}{the ^y score of the model}

\item{y:}{the response vector}

\item{verbose:}{print running information when set to TRUE}

\item{sign:}{weather the score should be greater or smaller than the intercept (default:"auto")}

\item{return.all:}{if TRUE, the function will return the intercept as well as the table used to compute it.}

\item{plot:}{if TRUE, the score will be visialized (default:FALSE)}
}
\value{
the intercept, the sign and the accuracy
}
\description{
Computes the best intercept for the model
}
