% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga1.lib.R
\name{crossing}
\alias{crossing}
\title{Creates new combinations of features based from a parents.}
\usage{
crossing(clf, pop, parents, seed = NULL)
}
\arguments{
\item{clf:}{the classifier parameter object}

\item{pop:}{A population (i.e. list) of index vectors}

\item{parents:}{Indexes of the population pointing to the subset of the population containing the parents (whose genes/features) will be used to create the children.}

\item{seed:}{For reproductibility purpose to fix the random generator number.}
}
\value{
a population of models, containing parents and children
}
\description{
This function is used in terga1 will create new combinations of features based of existing ones from the parents.
}
