% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{denseVecToModel}
\alias{denseVecToModel}
\title{denseVecToModel}
\usage{
denseVecToModel(X, y, v, clf, eval.all = FALSE, obj = NULL)
}
\arguments{
\item{X:}{dataset}

\item{y:}{labels}

\item{v:}{A vector of coeffs (example v=c(0.0,1.0,0.0,-1.0))}

\item{clf:}{classifier information}

\item{eval.all:}{If TRUE the fitting of the function and intercept will be computed}

\item{obj:}{an object model to add to the model (default:NULL)}
}
\value{
an model object
}
\description{
Builds a model object based on model that is in the dense (long) format.
}
