% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{digest}
\alias{digest}
\title{Summarize the results from an experiment object}
\usage{
digest(
  obj,
  penalty = NULL,
  best.cv = TRUE,
  best.k = NULL,
  plot = FALSE,
  omit.na = TRUE
)
}
\arguments{
\item{obj:}{The experiment object resulting from the learning process `fit()`}

\item{penalty:}{A coefficient between 0 and 1, which is applied to penalize 
the performance of models as a consequence of model-size. We use this to select
the best model of the population of models (default:NULL)}

\item{best.cv:}{Should we chose the best model based on information learnerd 
cross validation (default:TRUE). This will work if the crossvalidation data is 
available. If not the best model will be selected with empirical results.}

\item{best.k:}{If we do not wish to let the algorithm select the model size, 
we can fix this by setting the best.k with an integer indicating the number of 
variables in the model (default:NULL).}

\item{plot:}{Should the digested results be plotted ? (default:FALSE)}

\item{omit.na:}{Omit data with empty results (default:TRUE)}
}
\value{
an object with digested information such as the best models for each 
model-size, their respective scores, the best model.
}
\description{
Sumarizes the results of an experiment object of the type 
`obj$classifier` and `obj$crossval`. This is different from the digestMC(),
which sumarizes a model collection obj$models
}
