% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{digestModelCollection}
\alias{digestModelCollection}
\title{digestModelCollection}
\usage{
digestModelCollection(obj, X = NULL, clf, k.penalty = 0, mmprev = FALSE)
}
\arguments{
\item{obj:}{a modelCollection object}

\item{X:}{the dataset (default = NULL)}

\item{clf:}{the classifier object}

\item{k.penalty:}{the penalty to apply for sparsity (default:0).}

\item{mmprev:}{activate the max.min.prevalence selector (default:FALSE)}
}
\value{
an object with sumarized results such as the best models for each k_sparse, their respective scores, the best model, etc
}
\description{
Sumarizes the results of a model collection object of the type clf_res$models. This is different from the digest() which sumarizes an experiment clf_res$classifier$models
}
\details{
Summarize the results from a given modelCollection object
}
