% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{disectModel}
\alias{disectModel}
\title{Analyzes the score construction and model}
\usage{
disectModel(mod, X, y, clf, plot = TRUE)
}
\arguments{
\item{mod:}{a model object where the score will be computed}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the class vector}

\item{clf:}{an object containing the different parameters of the classifier}

\item{plot:}{plot graphical interpretation of if TRUE, (default:TRUE)}
}
\value{
an object containing statistics on a given model
}
\description{
Analyzes the score construction and model
}
