% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{evaluateAdditionnalMetrics}
\alias{evaluateAdditionnalMetrics}
\title{Compute other prediction scores such as precision, recall and f-score}
\usage{
evaluateAdditionnalMetrics(mod, X, y, clf, mode = "train")
}
\arguments{
\item{mod:}{a model object to be evaluated}

\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}

\item{mode:}{training or testing mode}
}
\value{
a model whose evaluation parameters are updated
}
\description{
This function computes prediction scores based on the confusion matrix such as accuracy, precision, recall and f-score
}
