% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{evaluateAUC}
\alias{evaluateAUC}
\title{Computes the AUC of a model}
\usage{
evaluateAUC(score, y, sign = ">")
}
\arguments{
\item{score:}{the ^y score of the model}

\item{y:}{the response vector}

\item{sign:}{in which direction to make the comparison? "auto" (default): automatically define in which group 
the median is higher and take the direction accordingly. ">": if the predictor values for the control group 
are higher than the values of the case group (controls > t >= cases). "<": if the predictor values for the 
control group are lower or equal than the values of the case group (controls < t <= cases).}
}
\value{
an auc value
}
\description{
Computes the AUC of a model
}
