% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{evaluateModelRegression}
\alias{evaluateModelRegression}
\title{Evaluates the fitting coefficents of a model object}
\usage{
evaluateModelRegression(
  mod,
  X,
  y,
  clf,
  eval.all = FALSE,
  force.re.evaluation = FALSE
)
}
\arguments{
\item{mod:}{a model object}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{the classifier parameter object}

\item{eval.all:}{should the function evaluate all the scores (default:FALSE)}

\item{force.re.evaluation:}{re-evaluate all the scores even if they exist (default:FALSE)}
}
\value{
a model object with the fitting scores evaluated
}
\description{
Evaluates the fitting coefficients of a model object.
}
