% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{evaluatePopulation}
\alias{evaluatePopulation}
\title{evaluatePopulation}
\usage{
evaluatePopulation(
  X,
  y,
  clf,
  pop,
  eval.all = FALSE,
  force.re.evaluation = FALSE,
  estim.feat.importance = FALSE,
  mode = "train",
  delete.null.models = TRUE,
  lfolds = NULL
)
}
\arguments{
\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the class vector}

\item{clf:}{the object containing the classifier information}

\item{pop:}{the population of models to be evaluated}

\item{eval.all:}{should the function evaluate all the scores for each of the models (default:FALSE)}

\item{force.re.evaluation:}{re-evaluate all the scores even if they exist for each of the models (default:FALSE)}

\item{estim.feat.importance:}{evaluate the importance in the model object for each of the models (default:FALSE)}

\item{mode:}{A choice from c("train", "test") indicates wether we wish to learn the threthold 
of each of the models (default:"train") or not "test" for the c("terinter","bininter","ratio") languages}

\item{delete.null.models:}{should null indivuals be deleted (default:TRUE)}

\item{lfolds:}{compute evaluation in crossval (default:NULL)}
}
\value{
an individual object
}
\description{
Evaluates an entire population of models, that be predomics objects or individuals
}
