% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{evaluatePrevalence}
\alias{evaluatePrevalence}
\title{Evaluate the prevalence of a given model}
\usage{
evaluatePrevalence(mod, X)
}
\arguments{
\item{mod:}{a model object}

\item{X:}{dataset where to compute the prevalence}
}
\value{
A vector containing the prevalence of each feature
}
\description{
Evaluate the prevalence of a given model
}
