% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{evaluateYhat}
\alias{evaluateYhat}
\title{Computes the predected classification using a given model}
\usage{
evaluateYhat(
  mod = NULL,
  X,
  y,
  clf,
  score = NULL,
  intercept = NULL,
  sign = NULL
)
}
\arguments{
\item{mod:}{a model object to be used in the class prediction}

\item{X:}{dataset to classify}

\item{y:}{variable to predict}

\item{clf:}{an object containing the different parameters of the classifier}

\item{score:}{the score passed directly}

\item{intercept:}{the intercept passed directly}

\item{sign:}{the sign passed directly}
}
\value{
a vector with the predicted classification of the samples
}
\description{
This function evaluates the predicted classification either using (1) a model object that contains intercept and sign or (2) directly the attributes score, intercept, sign
}
