% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{evolve2m}
\alias{evolve2m}
\title{Second version of the evolve method}
\usage{
evolve2m(X, y, clf, pop, featEval, generation)
}
\arguments{
\item{X:}{Dataset to classify}

\item{y:}{Variable to predict}

\item{clf:}{The claffifier  object containing the different parameters}

\item{pop:}{The population to be evolved}

\item{featEval:}{A dataframe with the evaluation of each variable of the dataset, 
used for some mutator.}
}
\value{
A list with the size_pop bests of the combination of the old population 
with the new population
}
\description{
This evolve method realize the selection of the parents with 
two methods (for the moment) : elite and random. The it tags every selected 
individual with the index of it's mate in the population. Then the individual 
which sould be mutated are tagged. for each individual of the population we 
check if they need to be crossed and/or mutated and the we apply the operations 
and create a new individual for each operation applied.
}
