% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{filterFeaturesByPrevalence}
\alias{filterFeaturesByPrevalence}
\title{Selects the most prevalent features in the dataset baset on the provided thresholds.}
\usage{
filterFeaturesByPrevalence(
  X,
  y = NULL,
  nb.prevalence = NULL,
  perc.prevalence = NULL,
  by.class = TRUE
)
}
\arguments{
\item{X:}{the dataset X}

\item{y:}{the class vector (default:NULL)}

\item{nb.prevalence:}{the minimum number of non zero observations (default: 10)}

\item{perc.prevalence:}{the percentage of non zero observations (default: NULL)}

\item{by.class:}{wether the filter should be applied by class (default: TRUE)}
}
\value{
the filtered dataset, without the features that do not pass the filter.
}
\description{
Filters out all features that display a prevalence below a given threshold provided as a number of 
observations or percentage. This for the total dataset or by class.
}
