% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{filterfeaturesK}
\alias{filterfeaturesK}
\title{Selects a the top k features that are significantly associated with the class to predict}
\usage{
filterfeaturesK(
  data,
  trait,
  k = 10,
  type = "wilcoxon",
  restrict = rep(TRUE, ncol(data)),
  multiple.adjust = "BH",
  paired = FALSE,
  sort = TRUE,
  verbose = FALSE,
  verbose.step = NULL,
  return.data = FALSE
)
}
\arguments{
\item{data:}{the dataset X}

\item{trait:}{is the equivalent of y (class, or numerical)}

\item{k:}{the number of features (default:10)}

\item{type:}{the statistics to be run (default:wilcoxon)}

\item{restrict:}{Run the statistics in a subset of the dataset (default: a vector of all TRUE)}

\item{multiple.adjust:}{the multiple testing adjustment method (default:BH)}

\item{paired:}{wether paired statistics should be run (default:FALSE)}

\item{sort:}{return variables sorted by p-value significance (default:TRUE)}

\item{verbose:}{print out information indicating progress (default:FALSE)}

\item{verbose.step:}{Showing a 1 percent progress.}

\item{return.data:}{if (default:FALSE) this returns the statistics of X, otherwise the restricted data subset}
}
\description{
Runs statistics on the data and selects a subset of k features that are the most significant. 
Besides filtering this function can be used in a more larger statistical context.
}
