% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{filterNoSignal}
\alias{filterNoSignal}
\title{filterNoSignal: Omits the variables with no information}
\usage{
filterNoSignal(X, side = 1, threshold = "auto", verbose = FALSE)
}
\arguments{
\item{X:}{the dataset to clean}

\item{side:}{side=1 means that variables are in the rows. Other than 1 it will transpose the dataset}

\item{threshold:}{auto, will compute the first derivate of the median(sd)/x and will find an automatic threshold. When threshold is a numerical it will be used as a threshold and when is something else, will automatically be 0.}

\item{verbose:}{print out information when TRUE (default:FALSE)}
}
\description{
This function will clean a dataset from the variables that have no or little information.
}
