% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terda.lib.R
\name{findk}
\alias{findk}
\title{Find the number of weights not yet integer.}
\usage{
findk(
  x,
  y,
  nfolds = 1,
  gamma = 1,
  k.sparse = NULL,
  vartype = "real",
  lb = -1,
  ub = 1
)
}
\arguments{
\item{x}{matrix or dataframe of predictors, of dimension n*p; each row is an observation vector.}

\item{y}{response variable (1 or -1)}

\item{nfolds}{k-folds cross-validation that we want test. Dedault value is 1.}

\item{gamma}{is the hinge loss parameter.. Defines the margin}

\item{k.sparse}{is the sparsity (non-negative real value). Default value is \code{k.sparse = NULL} - no constraint.}

\item{vartype}{is the type of coefficients : \code{cplexAPI::CPX_INTEGER, cplexAPI::CPX_BINARY, cplexAPI::CPX_CONTINUOUS}. Default \code{vartype = cplexAPI::CPX_INTEGER}}

\item{lb}{is the lower bound of coefficients}

\item{ub}{is the upper bound of coefficients}
}
\value{
An integer is number of coefficients not yet integer.
}
\description{
This method return a maximum number of weights of the model not yet integer.
}
