% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terbeam.lib.R
\name{generateAllCombinations}
\alias{generateAllCombinations}
\title{generateAllCombinations}
\usage{
generateAllCombinations(X, y, clf, ind.features.to.keep, sparsity, allFeatures)
}
\description{
Generate every possible combination of a list of features and evaluate them
}
