% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{get_IndividualToBeMutated}
\alias{get_IndividualToBeMutated}
\title{Return list of individuals to mutate}
\usage{
get_IndividualToBeMutated(pop)
}
\arguments{
\item{pop:}{population list}
}
\value{
a list of individuals (= a population) containing all the individuals selected for mutation
}
\description{
Wraper function that returns the list of individuals that are going to go through 
the mutate function (normaly taged with \link[predomics]{tag_SelectRandom})
}
