% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{get_Parents}
\alias{get_Parents}
\title{Return list of parents}
\usage{
get_Parents(pop)
}
\arguments{
\item{pop:}{population list}
}
\value{
a list of individuals (= a population) containing all the selected parents
}
\description{
Wraper function that returns the list of parents (normaly taged with 
\link[predomics]{tag_SelectRandom})
}
