% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getFeaturePrevalence}
\alias{getFeaturePrevalence}
\title{Evaluates the prevalence of a list of features in the whole dataset and per each class}
\usage{
getFeaturePrevalence(features, X, y = NULL, prop = TRUE, zero.value = 0)
}
\arguments{
\item{features:}{a list of features or features indexes for which we wish to compute prevalence}

\item{X:}{dataset where to compute the prevalence}

\item{y:}{if provided it will also compute hte prevalence per each class (default:NULL)}

\item{prop:}{weather to compute the prevalence in number or as a proportion (default:TRUE)}

\item{zero.value:}{the value that specifies what is zero. This can be a different than 0 in log transformed data for instance (default = 0)}
}
\value{
A list containing the prevalence in the whole dataset as well as classes (if provided)
}
\description{
Evaluate the prevalence of a given model
}
