% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getFitIndividual}
\alias{getFitIndividual}
\title{Get the fitting score of an individual object}
\usage{
getFitIndividual(individual)
}
\arguments{
\item{individual}{: an individual object}
}
\value{
a fitting score
}
\description{
Get the fitting score of an individual object.
}
