% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getFitModel}
\alias{getFitModel}
\title{Get the fitting score of a model object}
\usage{
getFitModel(mod)
}
\arguments{
\item{mod}{: a model object}
}
\value{
a fitting score
}
\description{
Get the fitting score of a model object.
}
