% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getFitModels}
\alias{getFitModels}
\title{Get the fitting score of a list a models}
\usage{
getFitModels(pop)
}
\arguments{
\item{pop}{: a list of models}
}
\value{
a vector of fitting scores
}
\description{
Get the fitting score of a list a models.
}
