% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getFitPopulation}
\alias{getFitPopulation}
\title{Get the fitting score of a list of individuals}
\usage{
getFitPopulation(pop)
}
\arguments{
\item{pop}{: a list of individuals}
}
\value{
a vector of fitting scores
}
\description{
Get the fitting score of a list of individuals.
}
