% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.lib.R
\name{getGraph}
\alias{getGraph}
\title{getGraph}
\usage{
getGraph(mat, X, threshold = 0)
}
\arguments{
\item{X:}{dataset to classify}

\item{mat:}{AnalyseStableModels()$origin}

\item{threshold:}{Used to select a number of edgets. By default this is zero.}
}
\value{
a graph
}
\description{
This function gets a graph of the result of analyseStableModels
}
