% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getIndicesIndividual}
\alias{getIndicesIndividual}
\title{Get the index of the features in a given individual}
\usage{
getIndicesIndividual(individual)
}
\arguments{
\item{individual}{: an individual object}
}
\value{
the indices of the features
}
\description{
Get the indices of the features used in the individuals
}
