% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getIndicesPopulation}
\alias{getIndicesPopulation}
\title{Get the indices of the features used in a population of individuals}
\usage{
getIndicesPopulation(pop)
}
\arguments{
\item{pop}{: a list of individuals}
}
\value{
a matrix of indices (rows), and individuals (cols)
}
\description{
Get the indices of the features used in a population of individuals
}
