% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getMaxMinPrevalenceModel}
\alias{getMaxMinPrevalenceModel}
\title{Get the model that has the highest minimal prevalence in its features}
\usage{
getMaxMinPrevalenceModel(pop, X = NULL, evalToOrder = "fit_", selected = 0)
}
\arguments{
\item{pop:}{a population of model objects}

\item{X:}{dataset where to compute the prevalence}

\item{evalToOrder:}{which score should we use to order the models and select them (default:fit_)}

\item{selected:}{the number of selected models (default:0). If 0, everything is returned.}
}
\value{
a model or a list of model objects
}
\description{
Get the model that has the highest minimal prevalence in its features
}
