% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{getSign}
\alias{getSign}
\title{Evaluates the sign for a given feature this is the old getMgsVsTraitSignDiscr function}
\usage{
getSign(X, y, clf = NULL, parallel.local = FALSE)
}
\arguments{
\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{the classifier parameter object}

\item{parallel.local:}{weather or not to run in //}
}
\value{
a vector of +1 & -1 for each variable
}
\description{
Evaluates the sign for a given feature this is the old getMgsVsTraitSignDiscr function.
}
