% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibd}
\alias{ibd}
\title{Inflammatory Bowel Disease (frequencies) from the MetaHIT study}
\description{
This dataset consists of frequency abundance files as downloaded from http://waldronlab.io/curatedMetagenomicData/
This is a list containing two elements: (i) the X data matrix with 1045 species and 396 observations and (ii) patient class = -1 (n=148) and healthy controls (n=248)
}
\author{
Nielsen, H Bjørn, Mathieu Almeida, Agnieszka Sierakowska Juncker, Simon Rasmussen, Junhua Li, Shinichi Sunagawa, Damian R Plichta, et al “Identification and assembly of genomes and genetic elements in complex metagenomic samples without using reference genomes.” Nature biotechnology (July 6, 2014): 1–11.
}
\keyword{bowel}
\keyword{disease,}
\keyword{inflamatory}
\keyword{microbiome,}
\keyword{species}
