% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{index2names}
\alias{index2names}
\title{index2names}
\usage{
index2names(X, var.ind)
}
\arguments{
\item{X:}{the dataset}

\item{var.ind:}{the feature index vector}
}
\value{
the names of the features
}
\description{
Transforms feature indexes into feature names
}
