% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{individual}
\alias{individual}
\title{Creates an object individual}
\usage{
individual(
  X,
  y,
  clf,
  coeffs = NULL,
  ind = NULL,
  eval.all = FALSE,
  signs = NULL,
  obj = NULL,
  res_clf = NULL
)
}
\arguments{
\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the class vector}

\item{clf:}{the object containing the classifier information}

\item{ind:}{the indexes of the variables forming the individual could be null if we give the function a dense vector (via the coeff parameter) or if we also want to generate the individual}

\item{coeffs:}{the coefficients of the model, it could be a dense vector (in this case, ind need to be null), or it could be only the non zero values, or if it's null a new individual will be genrated}

\item{obj:}{an object to be incorporated in the model (default:NULL). We use this usually for SOTA.}

\item{res_clf:}{if provided information on mda etc can be found and transmitted to the model object}
}
\value{
an individual (model) object
}
\description{
Creates an object individual
}
