% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isClf}
\alias{isClf}
\title{Evaluates wether an object is a classifier}
\usage{
isClf(obj)
}
\arguments{
\item{obj:}{an object to test}
}
\value{
TRUE if the object is a classifier
}
\description{
Evaluates wether an object is a classifier
}
