% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isExperiment}
\alias{isExperiment}
\title{Evaluates wether an object is an experiment}
\usage{
isExperiment(obj)
}
\arguments{
\item{obj:}{an object to test}
}
\value{
TRUE if the object is an experiment
}
\description{
Evaluates wether an object is an experiment
}
