% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isLearnerSota}
\alias{isLearnerSota}
\title{Evaluates wether an object is a model SOTA SVM}
\usage{
isLearnerSota(obj)
}
\arguments{
\item{obj:}{a model to test}
}
\value{
TRUE if the object is a SOTA learner
}
\description{
Evaluates wether a learner is SOTA or not
}
