% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isModelBTR}
\alias{isModelBTR}
\title{Evaluates wether an object is a model BTR}
\usage{
isModelBTR(obj)
}
\arguments{
\item{obj:}{a model to test}
}
\value{
TRUE if the object is a model BTR
}
\description{
Evaluates wether an object is a model of type BTR
}
