% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isModelCollection}
\alias{isModelCollection}
\title{Evaluates wether an object is a model collection objecct}
\usage{
isModelCollection(obj)
}
\arguments{
\item{obj:}{an object to test}
}
\value{
TRUE if the object is a model collection objecct
}
\description{
Evaluates wether an object is a model collection objecct
}
