% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isModelSota}
\alias{isModelSota}
\title{Evaluates wether an object is a model SOTA}
\usage{
isModelSota(obj)
}
\arguments{
\item{obj:}{a model to test}
}
\value{
TRUE if the object is a model sota
}
\description{
Evaluates wether an object is a model of type sota
}
