% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isModelSotaGLMNET}
\alias{isModelSotaGLMNET}
\title{Evaluates wether an object is a model SOTA GLMNET}
\usage{
isModelSotaGLMNET(obj)
}
\arguments{
\item{mod:}{a model to test}
}
\value{
TRUE if the object is a model sota GLMNET
}
\description{
Evaluates wether an object is a model of type sota
}
