% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isModelSotaRF}
\alias{isModelSotaRF}
\title{Evaluates wether an object is a model SOTA RF}
\usage{
isModelSotaRF(obj)
}
\arguments{
\item{mod:}{a model to test}
}
\value{
TRUE if the object is a model sota RF
}
\description{
Evaluates wether an object is a model of type sota
}
