% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isModelSotaSVM}
\alias{isModelSotaSVM}
\title{Evaluates wether an object is a model SOTA SVM}
\usage{
isModelSotaSVM(obj)
}
\arguments{
\item{obj:}{a model to test}
}
\value{
TRUE if the object is a model sota SVM
}
\description{
Evaluates wether an object is a model of type sota
}
