% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isModelTerda}
\alias{isModelTerda}
\title{Evaluates wether an object is a model BTR Terda}
\usage{
isModelTerda(obj)
}
\arguments{
\item{mod:}{a model to test}
}
\value{
TRUE if the object is a model BTR Terda
}
\description{
Evaluates wether an object is a model of type BTR
}
