% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{isPopulation}
\alias{isPopulation}
\title{Evaluates wether an object is a population of models}
\usage{
isPopulation(obj)
}
\arguments{
\item{obj:}{an object to test}
}
\value{
TRUE if the object is a population
}
\description{
Evaluates wether an object is a population of models
}
