% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{listOfDenseVecToListOfModels}
\alias{listOfDenseVecToListOfModels}
\title{Builds a model object from a list of vector coefficients}
\usage{
listOfDenseVecToListOfModels(X, y, clf, v, lobj = NULL)
}
\arguments{
\item{X:}{dataset}

\item{y:}{labels}

\item{clf:}{classifier}

\item{v:}{list of vectors of coeffs. For example, v=list( c(0.0,1.0,0.0,-1.0) , c(1.0,1.0,0.0,0.0) , c(0.0,1.0,1.0,-1.0) )}

\item{lobj:}{a list of objects to add as elements in the model objects if not null (default:NULL)}
}
\value{
an model object
}
\description{
Builds a model object from a list of vector coefficients.
}
