% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{listOfDenseVecToModelCollection}
\alias{listOfDenseVecToModelCollection}
\title{Builds a list of dense vector coefficients from a list of models}
\usage{
listOfDenseVecToModelCollection(clf, X, y, v)
}
\arguments{
\item{clf:}{classifier}

\item{X:}{dataset}

\item{y:}{labels}

\item{v:}{list of dense vectors}
}
\value{
a model collection
}
\description{
Builds a list of dense vector coefficients from a list of models
}
