% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{listOfModels2ModelCollection}
\alias{listOfModels2ModelCollection}
\title{listOfModels2ModelCollection}
\usage{
listOfModels2ModelCollection(pop, nBest = NA)
}
\arguments{
\item{pop:}{is population (a list) of predomics objects}

\item{nBest:}{number of elements to return for each sparsity (default:NA)}
}
\value{
an model collection object
}
\description{
Structures a list of predomics objects into a structured collection by k_sparsity.
}
