% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{listOfModelsToListOfDenseVec}
\alias{listOfModelsToListOfDenseVec}
\title{Builds a list of dense vector coefficients from a list of models}
\usage{
listOfModelsToListOfDenseVec(clf, X, y, list.models)
}
\arguments{
\item{clf:}{classifier}

\item{X:}{dataset}

\item{y:}{labels}

\item{list.models:}{list of models}
}
\value{
a list of dense vectors of coefficient
}
\description{
Builds a list of dense vector coefficients from a list of models
}
