% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{listOfModelsToListOfSparseVec}
\alias{listOfModelsToListOfSparseVec}
\title{Builds a list of sparse vector coefficients from a list of models}
\usage{
listOfModelsToListOfSparseVec(list.models)
}
\arguments{
\item{list.models:}{list of models}
}
\value{
a list of dense vectors of coefficient
}
\description{
Builds a list of sparse vector coefficients from a list of models
}
