% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{loadPopulation}
\alias{loadPopulation}
\title{Load a population from a file}
\usage{
loadPopulation(fileName)
}
\arguments{
\item{fileName:}{The name of the file were the population is stored}
}
\value{
a population object
}
\description{
This function is used to load a population from a file on your disk (it must be in your working directory)
}
