% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.lib.R
\name{LPO_best_models}
\alias{LPO_best_models}
\title{Compute the cross-validation of leave one out for test stability}
\usage{
LPO_best_models(X, y, clf, p = 1, lfolds = NULL, return.all = FALSE, nk = 20)
}
\arguments{
\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{clf}

\item{lfolds:}{leave one out folds for cross-validation}

\item{return.all:}{return all results from the crossvalidation for feature stability testing}
}
\value{
a list containing generalisation scores for each fold as well as a matrix with the mean values.
}
\description{
Compute the cross-validation emprirical and generalization scores.
}
