% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{make.counter}
\alias{make.counter}
\title{Function used to create the counter for building clf$experiment$id}
\usage{
make.counter()
}
\description{
Function used to create the counter for building clf$experiment$id
}
