% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{makeFeatureAnnot}
\alias{makeFeatureAnnot}
\title{Prints as text the detail on a given experiment along with summarized results (if computed)}
\usage{
makeFeatureAnnot(pop, X, y, clf)
}
\arguments{
\item{pop:}{a population of models}

\item{X:}{the X dataset where to compute the abundance and prevalence}

\item{y:}{the target class}

\item{clf:}{an object containing the different parameters of the classifier}
}
\value{
a list with two data.frames one containing the coefficients per each model and the other a data.frame on the features
}
\description{
This function takes a population of models and creates a table with annotation on the features, 
such as prevalence in the models and dataset as well as different statistics
}
