% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{makeFeatureModelPrevalenceNetworkCooccur}
\alias{makeFeatureModelPrevalenceNetworkCooccur}
\title{Prints as text the detail on a given experiment along with summarized results (if computed)}
\usage{
makeFeatureModelPrevalenceNetworkCooccur(
  pop.noz,
  feature.annot,
  alpha = 0.05,
  verbose = TRUE,
  layout = "circlular"
)
}
\arguments{
\item{pop.noz:}{a data.frame of in features in the rows and models in the columns. 
This table contains the feature coefficients in the models and is obtained by makeFeatureAnnot()}

\item{feature.annot:}{a data frame with annotation on features obtained by makeFeatureAnnot()}

\item{alpha:}{the significane p-value of the co-occurance.}

\item{verbose:}{print out information during run}

\item{layout:}{the network layout by default is circular (layout_in_circle) and will be a weighted Fruchterman-Reingold otherwise}
}
\value{
plots a graph
}
\description{
This function will use the coocur package to compute the co-occurance of features in a population of models
}
