% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{mergeMeltBestScoreCV}
\alias{mergeMeltBestScoreCV}
\title{mergeMeltBestScoreCV}
\usage{
mergeMeltBestScoreCV(
  list.results.digest,
  k_catalogue = NULL,
  score = "auc_",
  penalty = 0,
  min.kfold.nb = FALSE
)
}
\arguments{
\item{list.results.digest:}{a list of digest objects one for each learner used. For example, list(res.terda.digest, res.terga.digest, res.terbeam.digest)}

\item{k_catalogue:}{the k_catalogue that will serve to build the result matrix (default:NULL)}

\item{score:}{the name of the score that needs to be used for the whole dataset visualization.}

\item{min.kfold.nb:}{wether we should restrict all experiments in the smallest number of k-folds of a comparative analyses (default = FALSE)}
}
\value{
a list of two data.frames
}
\description{
mergeMeltBestScoreCV returns a list of data frames that contain the best performance of the different learners without any focus on sparsity.
}
\details{
Merge a list of cross validation scores form digest results
}
