% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{mergeMeltImportanceCV}
\alias{mergeMeltImportanceCV}
\title{mergeMeltImportanceCV}
\usage{
mergeMeltImportanceCV(
  list.results,
  filter.cv.prev = 0.5,
  min.kfold.nb = FALSE,
  type = "mda",
  learner.grep.pattern = "*",
  nb.top.features = 25,
  feature.selection = NULL,
  fixed.order = FALSE,
  scaled.importance = TRUE,
  make.plot = TRUE,
  main = FALSE,
  cv.prevalence = TRUE
)
}
\arguments{
\item{list.results.digest:}{a list of digest objects one for each learner used. For example, list(res.terda.digest, res.terga.digest, res.terbeam.digest)}

\item{filter.cv.prev:}{filter variable for each learner based on the appearence prevalence in the cross validation.}

\item{min.kfold.nb:}{wether we should restrict all experiments in the smallest number of k-folds of a comparative analyses (default = FALSE)}

\item{type:}{the type of importance "mda (mean decreased accuracy)" or "pda (prevalence decreased accuracy)" (default = mda)}

\item{learner.grep.pattern:}{select a subset of learners using a grep pattern (default:"*")}

\item{nb.top.features:}{the number of top features to focus on the plot}

\item{feature.selection:}{the names of the features to be selected (default:NULL)}

\item{fixed.order:}{if the order of features in the plot should follow the feature selection one (default = FALSE)}

\item{scaled.importance:}{the scaled importance is the importance multipied by the prevalence in the folds. If (default = TRUE) this will be used, the mean mda 
will be scaled by the prevalence of the feature in the folds and ordered subsequently}

\item{make.plot:}{make a plot for all the learners}

\item{main:}{should add the title to the graph for correct alignment (default:FALSE)}

\item{cv.prevalence:}{wether or not to plot the distribution of the prevalence of the feature in the top-models for each k-fold in the graph (default:FALSE)}
}
\value{
a list of several data.frames and a ggplot object
}
\description{
mergeMeltImportanceCV returns a list of data frames that contain the feature importance of the different learners without any focus on sparsity.
}
\details{
Merge a list of cross validation scores form digest results
}
