% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{mergeMeltScoreCV}
\alias{mergeMeltScoreCV}
\title{mergeMeltScoreCV}
\usage{
mergeMeltScoreCV(
  list.results.digest,
  k_catalogue,
  generalization = TRUE,
  score = "auc_"
)
}
\arguments{
\item{list.results.digest:}{a list of digest objects one for each learner used. For example, list(res.terda.digest, res.terga.digest, res.terbeam.digest)}

\item{k_catalogue:}{the k_catalogue that will serve to build the result matrix}

\item{generalization:}{get results from CV generalization (if TRUE) or empirical otherwise (default: TRUE)}

\item{score:}{the name of the score that needs to be used for the whole dataset visualization.}
}
\value{
a list of two data.frames
}
\description{
mergeMeltScoreCV returns a list of data frames that contain the performance of each digest in the list with their sparsity.
}
\details{
Merge a list of cross validation scores form digest results
}
