% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{mergeResults}
\alias{mergeResults}
\title{mergeResults}
\usage{
mergeResults(
  list.results,
  sparsity = NULL,
  penalty = 0.001,
  best.k = NULL,
  colors = NULL,
  pch = NULL
)
}
\arguments{
\item{list.results:}{a list of Experiment objects one for each learner used. For example, list(res.terda, res.terga, res.terbeam)}

\item{sparsity:}{Sometimes a given method will have results with somehow different sparsity. This param will allow to set the catalogue of sparsity}

\item{best.k:}{a vector defining wether a given k should be used to set the best model selection (default:NULL).}

\item{colors:}{a vector defining the colors to be used in the graphics. If not specified they will be set by default. (default:NULL).}

\item{pch:}{a vector defining the shape of the points to be used in the graphics. If not specified they will be set by default. (default:NULL).}
}
\value{
list of data.frames and lists
}
\description{
mergeResults returns a list of data frames that contain the performance of each digest in the list with their sparsity.
}
\details{
Merge a list of Scores form a digest results
}
