% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{modelCollectionToPopulation}
\alias{modelCollectionToPopulation}
\title{Transform a model collection to a population (or list of model objects)}
\usage{
modelCollectionToPopulation(mod.collection)
}
\arguments{
\item{mod.collection:}{a modelCollection object organized by k_sparsity}
}
\description{
Transform a model collection to a population (or list of model objects)
}
