% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terda.lib.R
\name{multipleRR}
\alias{multipleRR}
\title{multipleRR}
\usage{
multipleRR(clf, X, y, w, n, remove.zero.vec = TRUE)
}
\arguments{
\item{clf:}{the classifier parameter object}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{w:}{a vector of wi coefficients}

\item{n:}{number of round roundings to compute}

\item{remove.zero.vec:}{whether to remove the zero vectors}
}
\value{
an population of dense vectors
}
\description{
computes multiple randomized rounding for a given vector of wi
}
