% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga1.lib.R
\name{mutate}
\alias{mutate}
\title{Changes feature indexes in a given percentage of models.}
\usage{
mutate(clf, pop, selection, seed = NULL)
}
\arguments{
\item{clf:}{the classifier parameter object}

\item{pop:}{A population (i.e. list) of index vectors}

\item{selection:}{Indexes of the population pointing to the subset of the models to be changed}

\item{seed:}{For reproductibility purpose to fix the random generator number.}
}
\value{
a population of models among which the mutated ones
}
\description{
This function is used in terga1 will create new combinations of features based of existing ones from the parents.
}
