% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minitools.R
\name{myAssert}
\alias{myAssert}
\title{Asserts a condition and prints a message or stops the block}
\usage{
myAssert(condition, message, stop = TRUE)
}
\arguments{
\item{condition:}{condition to be tested}

\item{message:}{message to be printed}

\item{stop:}{if TRUE stop the block}
}
\description{
Asserts a condition and prints a message or stops the block
}
