% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minitools.R
\name{myAssertNotNullNorNa}
\alias{myAssertNotNullNorNa}
\title{Asserts the existance of an object and prints a message or stops the block}
\usage{
myAssertNotNullNorNa(obj, message = "", stop = FALSE)
}
\arguments{
\item{obj:}{condition to be tested}

\item{message:}{message to be printed}

\item{stop:}{if TRUE stop the block}
}
\description{
Asserts the existance of an object and prints a message or stops
  the block
}
