% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{names2index}
\alias{names2index}
\title{names2index}
\usage{
names2index(X, var.names)
}
\arguments{
\item{X:}{the dataset}

\item{var.names:}{the feature names vector}
}
\value{
the index of the features
}
\description{
Transforms feature names feature indexes
}
