% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obesity}
\alias{obesity}
\title{Obesity (frequencies) from the MetaHIT study}
\description{
This dataset consists of frequency abundance files as downloaded from http://waldronlab.io/curatedMetagenomicData/
This is a list containing two elements: (i) the X data matrix with 1045 species and 292 observations and (ii) patient class = -1 (n=167) and healthy controls (n=96).
Caution, this dataset has also a class 0 with overweight patients, which needs to be omited from both X and y
}
\author{
Le Chatelier, Emmanuelle, Trine Nielsen, Junjie Qin, Edi Prifti, Falk Hildebrand, Gwen Falony, Mathieu Almeida, et al “Richness of human gut microbiome correlates with metabolic markers.” Nature 500, no. 7464 (April 9, 2014): 541–546.
}
\keyword{microbiome,}
\keyword{obesity,}
\keyword{species}
