% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotComparativeResults}
\alias{plotComparativeResults}
\title{Plot performance scores for multiple learners}
\usage{
plotComparativeResults(
  digested.results,
  plot = TRUE,
  ylim = c(0.5, 1),
  best = FALSE,
  ci = FALSE,
  main = "",
  mode = "classification"
)
}
\arguments{
\item{digested.results:}{a list of data.frames containing performance results from a lists of learners. This data object is returned by the function merge_digestScores()}

\item{ylim:}{y-axis zoom in the plot}

\item{best:}{a swith to plot the best values instead of declining by k_sparsity}

\item{main:}{name of the graphic}

\item{mode:}{either classification or regression (default:classification)}
}
\value{
A list of ggplot graphs if plot is set to FALSE and a pannel organized graph otherwise.
}
\description{
plotComparativeResults plots a digested.results data object to compare performance results between different learners.
}
