% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotComparativeResultsBest}
\alias{plotComparativeResultsBest}
\title{Plot performance scores for multiple learners}
\usage{
plotComparativeResultsBest(digested.results, plot = TRUE, ylim = c(0.5, 1))
}
\arguments{
\item{digested.results:}{a list of data.frames containing performance results from a lists of learners. This data object is returned by the function merge_digestScores()}

\item{ylim:}{y-axis zoom in the plot}
}
\value{
A list of ggplot graphs if plot is set to FALSE and a pannel organized graph otherwise.
}
\description{
plotComparativeResultsBest plots a digested.results data object to compare performance results between different learners focusing at the best model.
}
