% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotModel}
\alias{plotModel}
\title{Plots a model or a population of model objectsas barplots of scaled coefficients.}
\usage{
plotModel(
  mod,
  X,
  y,
  sort.features = FALSE,
  sort.ind = NULL,
  feature.name = FALSE,
  col.sign = c("deepskyblue1", "firebrick1"),
  main = "",
  slim = FALSE,
  importance = FALSE,
  res_clf = NULL
)
}
\arguments{
\item{mod:}{a model to plot}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the class vector}

\item{sort.features:}{wether the features need to be sorted by correlation with 'y' or not (default: TRUE)}

\item{sort.ind:}{computing sorting can take time if computed for every model and can be computed outside the function and passed as a parameter}

\item{feature.name:}{show the name of the features (default:FALSE)}

\item{col.sign:}{the colors of the cofficients based on the sign of the coefficients (default: -1=deepskyblue1, 1:firebrick1)}

\item{main:}{possibility to change the title of the function (default:"")}

\item{slim:}{plot without axis information (default:FALSE)}

\item{importance:}{the importance (mda) of the features in crossval}

\item{res_clf:}{the result of the learning process (default:NULL). If provided information on MDA will be extracted for the importance graphic.}
}
\description{
Plots a model or a population of models as a barplots, representing each feature, the length being the coefficient
}
