% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotPopulation}
\alias{plotPopulation}
\title{Plot Population of Models}
\usage{
plotPopulation(
  pop,
  X,
  y,
  sort.features = FALSE,
  sort.ind = NULL,
  col.sign = c("deepskyblue1", "firebrick1"),
  ncol = 10,
  slim = FALSE,
  importance = FALSE
)
}
\arguments{
\item{pop}{A population of models (either a list of models or a single
model).}

\item{X}{A data matrix or data frame where rows represent observations and
columns represent features.}

\item{y}{A vector of class labels (e.g., 1 and -1 for binary classification)
corresponding to the rows in `X`.}

\item{sort.features}{Logical; if `TRUE`, the features will be sorted by their
discriminance with respect to `y`. Default is `FALSE`.}

\item{sort.ind}{A vector of indices for sorting the features. If `NULL`, the
function will determine the order based on discriminance. Default is
`NULL`.}

\item{col.sign}{A vector of two colors (default is `c("deepskyblue1",
"firebrick1")`) used to represent positive and negative coefficients,
respectively.}

\item{ncol}{The number of columns to arrange the plots in (default is `10`).}

\item{slim}{Logical; if `TRUE`, the plots will be simplified. Default is
`FALSE`.}

\item{importance}{Logical; if `TRUE`, the feature importance will be plotted.
Default is `FALSE`.}
}
\value{
If the population consists of a single model, a plot of the model's
feature coefficients or importance is returned. If the population consists of
multiple models, a grid of plots is displayed using `grid.arrange`.
}
\description{
This function visualizes a population of models by plotting their feature
importance or coefficients. It supports sorting of features by discriminance
or importance, and allows for customization of colors and the number of
columns in the layout.
}
\details{
This function generates a series of plots for a population of
models, displaying the feature coefficients or importances. If the population
consists of multiple models, each model's coefficients or importances are
plotted in a separate subplot. The function supports feature sorting based on
their discriminance or importance, and customizes the layout of the plots.
}
\examples{
# Example usage for a population of models
pop <- list(model1, model2, model3)  # Assume these are pre-defined models
X <- data.frame(feature1 = rnorm(100), feature2 = rnorm(100))
y <- sample(c(1, -1), 100, replace = TRUE)

# Plot the population
plotPopulation(pop, X, y, sort.features = TRUE, ncol = 3)

}
\author{
Edi Prifti (IRD)
}
