% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{populationGet_X}
\alias{populationGet_X}
\title{Get the best model from a classifier result}
\usage{
populationGet_X(element2get, toVec = TRUE, na.rm = TRUE)
}
\arguments{
\item{element2get:}{the name of the attribute to get}

\item{toVec:}{should the results be unlisted (default:TRUE)}

\item{na.rm:}{delete the elements that are NA (default) when returning tovec}
}
\value{
a vector of attributes
}
\description{
Gets a given attribute from a population of predomics objects
}
