% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{populationSet_X}
\alias{populationSet_X}
\title{Set models with a given liist of objects}
\usage{
populationSet_X(pop, element2set = NULL, listwithelements = NULL)
}
\arguments{
\item{element2set:}{the name of the attribute to set}

\item{listwithelements:}{the list containing the elements to add}
}
\value{
an updated population
}
\description{
Sets a given attribute to the objects of the a given population
}
