% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{populationToDataFrame}
\alias{populationToDataFrame}
\title{populationToDataFrame}
\usage{
populationToDataFrame(
  pop,
  attributes = c("learner", "language", "fit_", "unpenalized_fit_", "auc_", "accuracy_",
    "cor_", "aic_", "intercept_", "eval.sparsity", "sign_", "precision_", "recall_",
    "f1_")
)
}
\arguments{
\item{pop:}{a list of model objects, (i.e a population of models)}

\item{attributes:}{the list of attributes that we wish to have in the data.frame (default:"learner","language","fit_", "unpenalized_fit_", "auc_", "accuracy_", "cor_", "aic_", "intercept_", "eval.sparsity", "sign_","precision_", "recall_","f1_")}
}
\value{
an data frame with attributes for each model
}
\description{
For each model in the list of models it will extract each attribute and create a dataframe needed for further exploration
}
