% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{predictPopulation}
\alias{predictPopulation}
\title{Predicts the outcomes of a population of models in a dataset.}
\usage{
predictPopulation(
  X,
  y,
  clf,
  pop,
  plot = FALSE,
  col = c("deepskyblue4", "firebrick4"),
  fill = c("deepskyblue1", "firebrick1")
)
}
\arguments{
\item{X:}{dataset to predit}

\item{y:}{the real y}

\item{pop:}{a population of models. This can be an FBM for instance.}

\item{clf:}{an object containing the different parameters of the classifier}

\item{plot:}{logical (default:FALSE), returns a barplot for each sample with 
the prevalence of each class by the population of models}

\item{col:}{manual scaling colors (default:`c("deepskyblue4", "firebrick4")`).}

\item{fill:}{manual scaling colors (default:`c("deepskyblue1", "firebrick1")`).}
}
\value{
A dataframe of predicted classes with samples in columns and models 
in rows, `if(plot = FALSE)`. Otherwise it will return a ggplot object.
}
\description{
This function evaluates the class of a population of models for a dataset X
}
