% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{printExperiment}
\alias{printExperiment}
\title{Prints as text the detail on a given Experiment object}
\usage{
printExperiment(obj, indent = "\\t--- ")
}
\arguments{
\item{obj:}{an Experiment object}

\item{indent:}{a string (default:'tab---') that will precede each element of the object.}
}
\value{
NULL if the object is not a valid Experiment
}
\description{
This function prints a summary of an Experiment object.
}
