% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{printModelCollection}
\alias{printModelCollection}
\title{Prints as text the detail on a given ModelCollection object}
\usage{
printModelCollection(obj, indent = "\\t--- ", method = "long")
}
\arguments{
\item{obj:}{a ModelCollection object}

\item{indent:}{a string (default:'tab---') that will precede each element of the object for the "long" method.}

\item{method:}{the output method (default:long) will print for each k_sparsity a short information of the population of models, 
while the short method will output the number of models for each k_sparsity}
}
\value{
NULL if the object is not a valid ModelCollection.
}
\description{
This function prints a ModelCollection object. For each k_sparsity it will show some detail of 
the maximum first models
}
