% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{printy}
\alias{printy}
\title{Prints as text the detail on a given object from the predomics package.}
\usage{
printy(obj)
}
\arguments{
\item{obj:}{an object from the predomics object}
}
\description{
This function will summarize any of the predomics package objects such as can be an Experiment, 
a Model, a Population of models or a ModelCollection
}
