% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{resetTags}
\alias{resetTags}
\title{Resets selection, mutation and mate tags to inactive}
\usage{
resetTags(pop, selected = FALSE, toBeMutated = FALSE, mate = -1)
}
\arguments{
\item{pop:}{The population to be evolved}

\item{selected:}{set to (default:FALSE) the selected attribute, if not null.}

\item{toBeMutated:}{set to (default:FALSE) the selected attribute, if not null.}

\item{mate:}{set to (default:-1) the selected attribute, if not null.}
}
\value{
A modified population
}
\description{
Resets selection, mutation and mate tags to inactive
}
