% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predomics.R
\name{runClassifier}
\alias{runClassifier}
\title{Runs the learning on a dataset}
\usage{
runClassifier(X, y, clf, x_test = NULL, y_test = NULL)
}
\arguments{
\item{X:}{The dataset to classify}

\item{y:}{The variable to predict}

\item{clf:}{The classifier object containing the different settings of the classifier.}

\item{x_test:}{if not NULL (default) this dataset will be used to evaluate the models in a subset for the feature importance}

\item{y_test:}{if not NULL (default) this dataset will be used to evaluate the models in a subset for the feature importance}
}
\value{
the classifier along with the classification results as a sub-element
}
\description{
This function runs a classifier in a given dataset
}
