% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{savePopulation}
\alias{savePopulation}
\title{Save a population to a file}
\usage{
savePopulation(pop, fileName, compress = TRUE)
}
\arguments{
\item{pop:}{The population to be saved}

\item{fileName:}{The name of the file were you want to save the population}
}
\description{
You can use this function to save a population to a file on you're disk (it will be in your working directory)
}
