% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{scoreRatio}
\alias{scoreRatio}
\title{Computes the ^y score of the model as a ratio}
\usage{
scoreRatio(class_1_score, class_2_score, epsilon = NULL)
}
\arguments{
\item{class_1_score:}{the sum score for the features of class 1}

\item{class_2_score:}{the sum score for the features of class 2}

\item{epsilon:}{is a very small value that will would avoid Inf values in the ratio. This can be either specified in the when setting the classifier and if not specified will be set as the minimum number of the machine (e.g. 2.23e-308). Caution this should be adapted when working with other types of data.}
}
\value{
a vector containing the predicted ^y score for each observation
}
\description{
Computes the ^y score of the model as a ratio
}
