% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{selectBestPopulation}
\alias{selectBestPopulation}
\title{Select the top significant best part of the population}
\usage{
selectBestPopulation(pop, score = "fit_", p = 0.05, k_penalty = 0, k_max = 0)
}
\arguments{
\item{pop:}{a list of model objects}

\item{score:}{the attribute of the model to be used for the evaluation}

\item{p:}{the p-value threshold}

\item{k_penalty:}{the penalty to apply to the score based on the k_sparsity (default:0)}

\item{k_max:}{select the best population below a given threshold. If (default:0) no selection is performed.}
}
\value{
a sub part of the population
}
\description{
This function allows to select the best part of a population that is significantly not different from the best model
}
