% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{selector_v1}
\alias{selector_v1}
\title{Does an elite selection on a population}
\usage{
selector_v1(pop, number, clf)
}
\description{
This function is a template for other selectors, it takes a population and a number of individuals to select. The result is the \code{number} bests element of the population.
}
