% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{selector_v2}
\alias{selector_v2}
\title{Randomly Select Individuals from a Population}
\usage{
selector_v2(pop, number, clf = NULL)
}
\arguments{
\item{pop}{A list representing the population from which individuals are
selected.}

\item{number}{An integer specifying the number of individuals to select
randomly from the population.}

\item{clf}{An optional classifier object that can contain additional
parameters for selection (not directly used in this function).}
}
\value{
A list of randomly selected individuals from the population.
}
\description{
This function randomly selects a specified number of individuals from a
population without considering their fitness scores. It is a simple selection
strategy that introduces diversity by choosing individuals at random.
}
\details{
The function selects `number` individuals randomly from the
population `pop`, without replacement. This approach does not prioritize
fitness and can be useful for introducing genetic diversity in evolutionary
algorithms.
}
\examples{
\dontrun{
pop <- list(
  list(fit_ = 0.9, indices_ = 1:5),
  list(fit_ = 0.8, indices_ = 6:10),
  list(fit_ = 0.7, indices_ = 11:15)
)
selected_individuals <- selector_v2(pop, number = 2)
print(selected_individuals)
}

}
