% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.lib.R
\name{sim_inter}
\alias{sim_inter}
\title{compare stability of different modeles (inter k)}
\usage{
sim_inter(tmp, X)
}
\arguments{
\item{X:}{dataset to classify}

\item{tmp:}{the digested result from digest}
}
\value{
a num
}
\description{
This function compares stability of different modeles (inter k)
}
