% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{sortPopulation}
\alias{sortPopulation}
\title{sortPopulation}
\usage{
sortPopulation(pop, evalToOrder = "fit_", decreasing = TRUE)
}
\arguments{
\item{pop:}{a population (list) of evaluated predomics objects}

\item{evalToOrder:}{the attribute to be used in the sorting (default:fit_)}

\item{decreasing:}{whether the sorting should be be decreasing or not (default:decreasing)}
}
\value{
a sorted population of predomics objects
}
\description{
Sort a population according to a given attribute (evalToOrder)
}
