% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota.glmnet.R
\name{sota.glmnet}
\alias{sota.glmnet}
\title{sota.glmnet}
\usage{
sota.glmnet(...)
}
\value{
an object containing a list of parameters for this classifier
}
\description{
sota.glmnet herits from terda and does not use the randomized rounding, using thus only the glmnet component
}
\details{
sota.glmnet: sota.glmnet classifier parameter function
}
