% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sota.svm.R
\name{sota.svm}
\alias{sota.svm}
\title{sota.svm}
\usage{
sota.svm(
  sparsity = c(1:30),
  objective = "auc",
  max.nb.features = 1000,
  intercept = 0,
  language = "svm",
  evalToFit = "auc_",
  k_penalty = 0,
  scaled = TRUE,
  type = NULL,
  kernel = "rbfdot",
  kpar = "automatic",
  C = c(1e-04, 0.001, 0.01, 0.1, 1, 10, 100, 1000, 10000),
  nu = 0.2,
  epsilon.hp = 0.1,
  prob.model = FALSE,
  class.weights = NULL,
  fit = TRUE,
  cache = 40,
  tol = 0.001,
  shrinking = TRUE,
  na.action = na.omit,
  popSaveFile = "NULL",
  seed = "NULL",
  nCores = 4,
  verbose = TRUE,
  plot = FALSE,
  warnings = FALSE,
  debug = FALSE,
  print_ind_method = "short",
  experiment.id = NULL,
  experiment.description = NULL,
  experiment.save = "nothing"
)
}
\arguments{
\item{language}{is the language that is used by the different algorithms {bin, bininter, ter, terinter, ratio}, (default:"sota")}

\item{sparsity:}{number of features in a given model. This is a vector with multiple lengths.}

\item{objective:}{prediction mode (default: auc)}

\item{max.nb.features:}{create the glmnet object using only the top most significant features (default:1000)}

\item{intercept:}{(Interceot for the a given model) (default:NULL)}

\item{evalToFit:}{Which model property will be used to select the best model among different k_sparsities (default: auc_)}

\item{k_penalty:}{Penalization of the fit by the k_sparsity (default: 0)}

\item{scaled:}{??}

\item{type:}{??}

\item{kernel:}{??}

\item{kpar:}{??}

\item{C:}{(??)}

\item{nu:}{??}

\item{epsilon.hp:}{(??) (for the SVM)}

\item{prob.model:}{??}

\item{class.weights:}{??}

\item{fit:}{??}

\item{cache:}{(??)}

\item{tol:}{??}

\item{shrinking:}{??}

\item{na.action:}{??}

\item{popSaveFile:}{(??)}

\item{seed:}{the seed to be used for reproductibility. If seed=NULL than it is not taken into account (default:NULL).}

\item{nCores:}{the number of CPUs to run the program in parallel}

\item{plot:}{Plot graphics indicating the evolution of the simulation (default:FALSE)}

\item{verbose:}{print out information on the progress of the algorithm (default:TRUE)}

\item{warnings:}{Print out warnings when runnig (default:FALSE).}

\item{debug:}{print out information on the progress of the algorithm (default:FALSE)}

\item{print_ind_method:}{One of c("short","graphical") indicates how to print a model and subsequently a population during the run (default:"short").}

\item{experiment.id:}{The id of the experiment that is to be used in the plots and comparitive analyses (default is the learner's name, when not specified)}

\item{experiment.description:}{A longer description of the experiment. This is important when many experiments are run and can also be printed in by the printExperiment function.}

\item{experiment.save:}{Data from an experiment can be saved with different levels of completness, with options to be selected from c("nothing", "minimal", "full"), default is "minimal"}
}
\value{
an object containing a list of parameters for this classifier
}
\description{
sota.svm is a wrapper that executes svm using the same framework as for the predomics package.
}
\details{
sota.svm: launching svm classifier
}
