% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{sparseVecToModel}
\alias{sparseVecToModel}
\title{sparseVecToModel}
\usage{
sparseVecToModel(X, y, v, clf, eval.all = FALSE, obj = NULL)
}
\arguments{
\item{X:}{dataset}

\item{y:}{labels}

\item{v:}{A vector of indexes (example v=c(1,11))}

\item{clf:}{classifier information}

\item{eval.all:}{Should the model be evaluated (default:FALSE)}

\item{obj:}{an object model to add to the model (default:NULL)}
}
\value{
an model object
}
\description{
Builds a model object based on model that is in the sparse (short) format.
}
