% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.lib.R
\name{summarySE}
\alias{summarySE}
\title{Plot performance scores for multiple learners.}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data:}{a data frame}

\item{groupvars:}{a vector containing names of columns that contain grouping variables}

\item{na.rm:}{a boolean that indicates whether to ignore NA's}

\item{conf.interval:}{the percent range of the confidence interval (default is 95\%)}
}
\value{
A transformed data frame with information on the different errors and confidence.
}
\description{
summarySE gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95\%).
}
