% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t2d}
\alias{t2d}
\title{Type 2 diabetes (frequencies) BGI}
\description{
This dataset consists of frequency abundance files as downloaded from http://waldronlab.io/curatedMetagenomicData/
This is a list containing two elements: (i) the X data matrix with 1045 species and 344 observations and (ii) patient class = -1 (n=170) and healthy controls (n=174)
}
\author{
Qin, Junjie, Yingrui Li, Zhiming Cai, Shenghui Li, Jianfeng Zhu, Fan Zhang, Suisha Liang, et al “A metagenome-wide association study of gut microbiota in type 2 diabetes.” Nature (September 26, 2012).
}
\keyword{2}
\keyword{diabetes,}
\keyword{microbiome,}
\keyword{species}
\keyword{type}
