% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t2dw}
\alias{t2dw}
\title{Type 2 diabetes (frequencies) Women Sweden}
\description{
This dataset consists of frequency abundance files as downloaded from http://waldronlab.io/curatedMetagenomicData/
This is a list containing two elements: (i) the X data matrix with 1045 species and 145 observations and (ii) patient class = -1 (n=53) and healthy controls (n=43)
Caution, this dataset has also a class 0 with IG patients, which needs to be omited from both X and y
}
\author{
Karlsson, Fredrik H, Valentina Tremaroli, Intawat Nookaew, Göran Bergström, Carl Johan Behre, Björn Fagerberg, Jens Nielsen, and Fredrik Bäckhed. “Gut metagenome in European women with normal, impaired and diabetic glucose control.” Nature (May 29, 2013): 1–7.
}
\keyword{2}
\keyword{diabetes,}
\keyword{microbiome,}
\keyword{species}
\keyword{type}
