% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{tag_Couples}
\alias{tag_Couples}
\title{Tag the couples}
\usage{
tag_Couples(pop, parents)
}
\arguments{
\item{pop:}{The population on which we the couples are being constituted;}

\item{parents:}{The parent candidate individuals from which the couples will 
be selected.}
}
\value{
The parent population with the couple tags set.
}
\description{
This function selects constitutes the couples that will give the
next generation individuals by adding the couple id on the mate attribute.
}
