% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga1.R
\name{terga1}
\alias{terga1}
\title{terga1}
\usage{
terga1(
  sparsity = c(1:10),
  size_pop = 100,
  size_world = "NULL",
  max.nb.features = 1000,
  popSourceFile = "NULL",
  popSaveFile = "NULL",
  language = "terinter",
  scoreFormula = scoreRatio,
  epsilon = "NULL",
  unique_vars = FALSE,
  objective = "auc",
  k_penalty = 0,
  evalToFit = "fit_",
  estimate_coefs = FALSE,
  intercept = "NULL",
  select_type = "mixed",
  select_perc1 = 20,
  select_perc2 = 30,
  perc_best_ancestor = 10,
  mutate_size = 70,
  mutate_rate = 50,
  nb_generations = 100,
  convergence = TRUE,
  convergence_steps = 10,
  evolve_k1 = TRUE,
  plot = FALSE,
  verbose = TRUE,
  warnings = FALSE,
  debug = FALSE,
  print_ind_method = "short",
  parallelize.folds = TRUE,
  nCores = 4,
  seed = "NULL",
  experiment.id = "NULL",
  experiment.description = "NULL",
  experiment.save = "nothing"
)
}
\arguments{
\item{language}{is the language that is used by the different algorithms {bin, bininter, ter, terinter, ratio}, (default:"terinter")}

\item{sparsity:}{number of features in a given model. This is a vector with multiple lengths.}

\item{size_pop:}{the number of individuals in a population to be evolved.}

\item{size_world:}{this is the number of features in the dataset.}

\item{max.nb.features:}{focuses only on the subset of top most significant features (default:1000)}

\item{popSourceFile:}{A population of models that can start as a first generation to be evolved (default:NULL).}

\item{popSaveFile:}{(??)}

\item{scoreFormula:}{a Function that contains the ratio Formula or other specific ones}

\item{epsilon:}{a small value to be used with the ratio language (default: NULL). When null it is going to be calculated by the minimum value of X divided by 10.}

\item{unique_vars:}{logical (default: FALSE) indicates weather unique variables can be used in a model or population.}

\item{objective:}{this can be auc, cor or aic. Terga can also predict regression, other than class prediction. (default:auc)}

\item{estimate_coefs:}{non ternary solution for the aic objective (default:FALSE)}

\item{intercept:}{(Interceot for the a given model) (default:NULL)}

\item{evalToFit:}{The model performance attribute to use as fitting score (default:"fit_"). Other choices are c("auc_","accuracy_","precision_","recall_","f_score_")}

\item{k_penalty:}{Penalization of the fit by the k_sparsity (default: 0)}

\item{select_type:}{the selection operator type. can be mixed, elite or tournoi (default: mixed)}

\item{select_perc1:}{percentage of individuals to be selected with elite}

\item{select_perc2:}{percentage of individuals to be selected with tournoi}

\item{perc_best_ancestor:}{percentage of best ancentors as seeding in the new population}

\item{mutate_size:}{percentage of individuals in the population to be mutated}

\item{mutate_rate:}{percentage of features in an individual to be mutated}

\item{plot:}{plot graphics indicating the evolution of the simulation (default:FALSE)}

\item{convergence:}{should the algorithm converge when the best individual is not improving (default:TRUE).}

\item{convergence_steps:}{the number of generations after which we consider convergence (default:10).}

\item{evolve_k1:}{weather or not to evaluate exhaustively the features for k_sparse=1. This will take a lot of time if the dataset is large, thus the possibility to evolve this using the GA. (default:TRUE)}

\item{verbose:}{print out information on the progress of the algorithm (default:TRUE)}

\item{warnings:}{Print out warnings when runnig (default:FALSE).}

\item{debug:}{print debug information (default:FALSE)}

\item{print_ind_method:}{One of c("short","graphical") indicates how to print a model and subsequently a population during the run (default:"short").}

\item{parallelize.folds:}{parallelize folds when cross-validating (default:TRUE)}

\item{nb_generations:}{maximum number of generations to evolve the population.}

\item{nCores:}{the number of cores to execute the program. If nCores=1 than the program runs in a non parallel mode}

\item{seed:}{the seed to be used for reproductibility. If seed=NULL than it is not taken into account (default:NULL).}

\item{experiment.id:}{The id of the experiment that is to be used in the plots and comparitive analyses (default is the learner's name, when not specified)}

\item{experiment.description:}{A longer description of the experiment. This is important when many experiments are run and can also be printed in by the printExperiment function.}

\item{experiment.save:}{Data from an experiment can be saved with different levels of completness, with options to be selected from c("nothing", "minimal", "full"), default is "minimal"}
}
\value{
an object containing a list of parameters for this classifier
}
\description{
terga1 is a model search algorithm based on genetic algorithms (GA). A “genome” or “individual” in this context is a combination of features that will be associated together to compute a score that will be the prediction model. Depending on the type of fitting function that is maximized the fatures are weighed by specific coefficients. In short the algorithm is based on different operations such as crossing, mutating and evolving different “individuals” and evaluating their fitness to the “environment” which is represented by the variable to be predicted.
}
\details{
terga1: Model search algorithm based on genetic algorithms (GA)
}
