% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeImportantFeaturesFBM.R
\name{analyzeImportanceFeaturesFBM}
\alias{analyzeImportanceFeaturesFBM}
\title{Visualize a summary of an experiment/set of experiments}
\usage{
analyzeImportanceFeaturesFBM(
  clf_res,
  X,
  y,
  makeplot = TRUE,
  saveplotobj = TRUE,
  name = "",
  verbose = TRUE,
  pdf.dims = c(width = 25, height = 20),
  filter.cv.prev = 0.25,
  nb.top.features = 100,
  scaled.importance = FALSE,
  k_penalty = 0.75/100,
  k_max = 0
)
}
\arguments{
\item{clf_res}{The result of an experiment or multiple experiments 
(list of experiments)}

\item{X}{The feature table used as input of fit function behind experiments 
in clf_res}

\item{y}{The target class (binary/continuous)}

\item{makeplot}{make a pdf file with the resulting plots (default:TRUE)}

\item{saveplotobj}{make a .Rda file with a list of the individual plots 
(default:TRUE)}

\item{name}{the suffix of the pdf file (default:"")}

\item{verbose}{print out informaiton}

\item{pdf.dims}{dimensions of the pdf object (default: c(w = 25, h = 20))}

\item{filter.cv.prev}{keep only features found in at least (default: 0.25, 
i.e 25 percent) of the cross validation experiments}

\item{nb.top.features}{the maximum number (default: 100) of most important 
features to be shown.
If the number of features in FBM < nb.top.features, the number of features 
in FBM will be shown instead}

\item{scaled.importance}{the scaled importance is the importance multiplied 
by the prevalence in the folds. If (default = TRUE) this will be used, the 
mean mda will be scaled by the prevalence of the feature in the folds and 
ordered subsequently}

\item{k_penalty}{the sparsity penalty needed to select the best models of 
the population (default:0.75/100).}

\item{k_max}{select the best population below a given threshold. 
If (default:0) no selection is performed.}
}
\value{
plots if makeplot is FALSE; plot.list list object saved locally with 
individual plots (including source data) if saveplotobj
}
\description{
Visualization of 4 panels corresponding to feature prevalence 
in FBM, feature importance, feature prevalence in groups, effect sizes of 
feature abundances vs y-variable (cliff's delta for binary y; 
spearman rho for continuous y). Can be applied to single classification task 
or to multiple classification tasks carried out on the same X-y dataset
}
