% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga1.lib.R
\name{evolve}
\alias{evolve}
\title{Creates new combinations of features based from a parents.}
\usage{
evolve(X, y, clf, pop, seed = NULL)
}
\arguments{
\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the response vector}

\item{clf:}{the classifier parameter object}

\item{pop:}{A population (i.e. list) of index vectors}

\item{seed:}{For reproductibility purpose to fix the random generator number.}
}
\value{
a population of models, containing parents and children
}
\description{
This function is used in terga1 and is the main engine of the algorithm that allows to cross, mutate and select individuals from one generation to the next.
}
