% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{get_IndividualToBeMutated}
\alias{get_IndividualToBeMutated}
\title{Retrieve Individuals Marked for Mutation}
\usage{
get_IndividualToBeMutated(pop)
}
\arguments{
\item{pop}{A list representing the population, where each individual has a
`toBeMutated` attribute indicating if they are marked for mutation.}
}
\value{
A vector of indices corresponding to individuals in the population
  who are marked for mutation.
}
\description{
This function identifies and returns the indices of individuals in the
population who are marked for mutation, as indicated by the `toBeMutated`
attribute.
}
\details{
The function uses `populationGet_X` to retrieve individuals with the
`toBeMutated` attribute set to `TRUE`. It returns the indices of these
individuals, allowing them to be processed for mutation in subsequent
evolutionary steps.
}
\examples{
\dontrun{
pop <- list(
  list(fit_ = 0.9, toBeMutated = TRUE),
  list(fit_ = 0.8, toBeMutated = FALSE),
  list(fit_ = 0.7, toBeMutated = TRUE)
)
toBeMutated <- get_IndividualToBeMutated(pop)
print(toBeMutated) # Should return indices of individuals marked for mutation
}

}
