% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotAbundanceByClass}
\alias{plotAbundanceByClass}
\title{Plots the prevalence of a list of features in the whole dataset and per each class}
\usage{
plotAbundanceByClass(
  features,
  X,
  y,
  topdown = TRUE,
  main = "",
  plot = TRUE,
  col.pt = c("deepskyblue4", "firebrick4"),
  col.bg = c("deepskyblue1", "firebrick1")
)
}
\arguments{
\item{features:}{a list of features or features indexes for which we wish to compute prevalence}

\item{X:}{dataset where to compute the prevalence}

\item{y:}{if provided it will also compute hte prevalence per each class (default:NULL)}

\item{topdown:}{showing features from top-down or the other way around (default:TRUE)}

\item{main:}{main title (default:none)}

\item{plot:}{if TRUE this provides a plot, otherwise will return different metrics such as prevalence and enrichment statistics}

\item{col.pt:}{colors for the point border (-1:deepskyblue4, 1:firebrick4)}

\item{col.bg:}{colors for the point fill (-1:deepskyblue1, 1:firebrick1)}
}
\value{
a ggplot object
}
\description{
Plots the abundance of a given number of features for each class and tests significance
}
