% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotAUCg}
\alias{plotAUCg}
\title{Plot the AUC of a given classifier}
\usage{
plotAUCg(mod = NULL, score, y, main = "", ci = TRUE, show.intercept = TRUE)
}
\arguments{
\item{mod:}{a predomics model object (default = NULL)}

\item{score:}{this is the y^ of a given model}

\item{y:}{the class to be predicted}

\item{main:}{title of the graph}

\item{ci:}{the point shape for the graph}

\item{show.intercept:}{plot or not the intercept on the graph (default:TRUE)}
}
\value{
a ggplot object
}
\description{
Analyze the results from a given classifier.
}
