% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotComparativeEmpiricalScore}
\alias{plotComparativeEmpiricalScore}
\title{Plots a graph for a given score}
\usage{
plotComparativeEmpiricalScore(
  digested.results,
  ylim = c(0.5, 1),
  score = "auc_",
  main = ""
)
}
\arguments{
\item{digested.results:}{a list of data.frames containing performance results from a lists of learners. This data object is returned by the function merge_digestScores()}

\item{ylim:}{y-axis zoom in the plot}

\item{score:}{default (auc_) score}

\item{main:}{name of the graphic}
}
\value{
A ggplot graphs
}
\description{
plotComparativeEmpiricalScore plots a digested.results data object for a given score.
}
