% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeImportantFeaturesFBM.R
\name{plotImportanceFeaturesFBMobjects}
\alias{plotImportanceFeaturesFBMobjects}
\title{Visualize a list containing outouts of getImportanceFeaturesFBMobjects}
\usage{
plotImportanceFeaturesFBMobjects(
  FBMobjList,
  verbose = TRUE,
  nb.top.features = 100,
  makeplot = TRUE
)
}
\arguments{
\item{FBMobjList}{List of outputs of getImportanceFeaturesFBMobjects 
function (1 list per experiment to combine)}

\item{verbose}{print out informaiton}

\item{nb.top.features}{features to retain for visualization (top features 
with highest mean feature importance across datasets)}

\item{makeplot}{make a pdf file with the resulting plots (default:TRUE)}
}
\value{
Combined visualization of feature prevalence in FBM + feature 
importance + feature effect size across groups + feature prevalence across 
groups in different predomics prediction tasks
}
\description{
Here we combine the 4 datasets generated by 
getImportanceFeaturesFBMobjects function from different prediction experiments 
(clf object + X + y) ; designed to combine predomics results with different 
X, y source data for unified visualization (feature prevalence in FBM, 
feature importance, feature effect size across groups, feature prevalence 
across groups)
}
