% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotModelScore}
\alias{plotModelScore}
\title{Plots a model or a population of model objectsas barplots of scaled coefficients.}
\usage{
plotModelScore(
  mod = NULL,
  y = NULL,
  col.sign = c("deepskyblue1", "firebrick1"),
  main = ""
)
}
\arguments{
\item{mod:}{a model to plot}

\item{y:}{the class to predict}

\item{col.sign:}{the colors of the cofficients based on the sign of the coefficients (default: -1=deepskyblue1, 1:firebrick1)}

\item{main:}{possibility to change the title of the function (default:"")}
}
\description{
Plots a model score or a population of models as a barplots, representing each feature, the length being the coefficient
}
