% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotPopulation}
\alias{plotPopulation}
\title{Plots a population of models (or a single model) objects as barplots of scaled coefficients.}
\usage{
plotPopulation(
  pop,
  X,
  y,
  sort.features = FALSE,
  sort.ind = NULL,
  col.sign = c("deepskyblue1", "firebrick1"),
  ncol = 10,
  slim = FALSE,
  importance = FALSE
)
}
\arguments{
\item{pop:}{a population of models to plot}

\item{X:}{the data matrix with variables in the rows and observations in the columns}

\item{y:}{the class vector}

\item{sort.features:}{wether the features need to be sorted by correlation with 'y' or not}

\item{sort.ind:}{computing sorting can take time if computed for every model and can be computed outside the function and passed as a parameter}

\item{col.sign:}{the colors of the cofficients based on the sign of the coefficients (default: -1=deepskyblue1,1:firebrick1)}

\item{ncol:}{number of graphics for each line (default: 10)}

\item{slim:}{plot without axis information (default:FALSE)}

\item{importance:}{the importance (mda) of the features in crossval}
}
\description{
Plots an model or a population of models as a barplots, representing each feature, the length being the coefficient
}
