% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{plotScoreBarcode}
\alias{plotScoreBarcode}
\title{Plots the barcode of the total score as well as positive and negative components}
\usage{
plotScoreBarcode(dscore, y, nb.col.levels = 30, main = "")
}
\arguments{
\item{dscore:}{an object containing different statistics on a model}

\item{y:}{the class vector}

\item{clf:}{an object containing the different parameters of the classifier}

\item{nb.col.levels:}{number of distinct colors from the viridis palette (default:30)}

\item{main:}{a title for the graphic}
}
\value{
nothing
}
\description{
Plots the barcode of the total score as well as positive and negative components
}
