% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga1.lib.R
\name{population}
\alias{population}
\title{Creates a population of index models.}
\usage{
population(
  clf,
  size_ind,
  size_world,
  best_ancestor = NULL,
  size_pop = NULL,
  seed = NULL
)
}
\arguments{
\item{clf:}{the classifier parameter object}

\item{size_ind:}{The sparsity of the models. All the models of this population will have the same number of features.}

\item{size_world:}{The number of features from which we can choose the indices. This is needed to compute the combinatory space search.}

\item{best_ancestor:}{We can supply to the popolution an individual (vector with indeces) of a lower sparsity. This will ensure to seed part of the population with at least those genes. We added this feature after an observations that a local optimum of lower sparsity was lost in higher sparsities.}

\item{size_pop:}{the number of models to produce (default=NULL). This information is stored here clf$params$size_pop, but this parameter allows to override it.}
}
\value{
a population of index models
}
\description{
This function is used in terga1 and generates a list of index vectors in the variable space. These vectors can be unique or not. NB that if clf$params$unique_vars is set to TRUE it can take a long time to come out of the while loop which ensures the uniqueness of the individuals.
}
