% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{printClassifier}
\alias{printClassifier}
\title{Prints as text the detail on a given Classifier object}
\usage{
printClassifier(obj, indent = "\\t--- ")
}
\arguments{
\item{obj:}{a Classifier object}

\item{indent:}{a string (default:'tab---') that will precede each element of the object.}
}
\value{
NULL if the object is not a valid Classifier
}
\description{
This function prints a summary of a Classifier object.
}
