% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.visu.R
\name{printModel}
\alias{printModel}
\title{# plot a horizontal barplot
#' @export
plotBarplot <- function(v, rev=TRUE, xlim=range(v), main=""){
  if(rev) v <- rev(v)
  barplot(v, las=2, horiz=TRUE, col="black", main=main, xlim=xlim)
}
Prints a model object as text.}
\usage{
printModel(mod, method = "short", score = "fit_")
}
\arguments{
\item{mod:}{a model to plot}

\item{method:}{an object containing the different parameters of the classifier}

\item{score:}{which score to show in the fit (default:fit_)}
}
\description{
Prints a model object as text
}
