% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{tag_select}
\alias{tag_select}
\title{Add `selected` tag using elite and random selection}
\usage{
tag_select(X, y, clf, pop)
}
\arguments{
\item{X:}{Unused but still here for compatibility}

\item{y:}{same as `X`}

\item{clf:}{the classifier object where parameters are defined}

\item{pop:}{the population on which we want to apply the selection}
}
\value{
the population with the tag `selected` on some of the individuals
}
\description{
This function combines \link[predomics]{tag_SelectElite} and \link[predomics]{tag_SelectRandom}
to tag the desired individuals in a population following the proportion given in the clf
}
