% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{tag_SelectElite}
\alias{tag_SelectElite}
\title{Tag individuals for parenting}
\usage{
tag_SelectElite(clf, pop, nbToSelect)
}
\arguments{
\item{clf:}{the classifier object}

\item{pop:}{the population on which we want to add the tag}

\item{nbToSelect:}{the number of individuals we are going to select in the population}
}
\value{
the population given as an input with `nbToSelect` bests individuals with `$selected = TRUE`
}
\description{
Function to add the tag "selected" to the best individuals of the population
}
