% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terga2.lib.R
\name{tag_ToBeMutated}
\alias{tag_ToBeMutated}
\title{Tag individuals for mutation}
\usage{
tag_ToBeMutated(pop, mutate_size, protected = NULL)
}
\arguments{
\item{pop:}{The population on which the individuals to be mutated will be selected}

\item{mutate_size:}{The number of individuals to mutate}

\item{protected:}{The index of individuals which should not be mutated.}
}
\value{
The population given as an input with `mutate_size` individuals with `toBeMutated = TRUE`
}
\description{
Function to add the tag "toBeMutated" to a randomly sampled part of 
the population. Some individuals (the best ones) will be protected from the mutation
so that genetic decline does not happen.
}
